Move test results examples to one file
